feat(follow-remote-merge): Check if remote branch exists before attempting to remove it
On the developer's main Gitlab repositories branches are removed when
branching through the UI. This caused an expected error message to
appear all the time.

Now the script checks if the remote branch exists before trying to
remove it. If it doesn't, everything is as expected, no action needs to
be taken and, subsequently, no warning needs to be printed.

**However**, if the remote branch still exists failing to remove it is
an error that needs to be properly alerted.

Both will happen now.