follow-remote-merge: Fix bug regarding missing remote
On branches that had no upstream target at all this helper would fail
with error. This shouldn't be the case, as this helper is intended to be
the one-stop solution to get rid of feature branches that were merged
already.

The commit shifts a bit of logic to fix this bug.